refactor: remove checks from main deploy script
