+++ packages/new-deploy/src/DeployContext.ts
@@ -149,10 +149,6 @@ export default class DeployContext extends MultiProvider<config.Domain> {
       ),
     );
 
-    // perform validation checks on core and bridges
-    await this.checkCores();
-    await this.checkBridges();
-
     return governanceTransactions;
   }
 
@@ -270,6 +266,12 @@ export default class DeployContext extends MultiProvider<config.Domain> {
         this.mustGetBridge(network).relinquish(),
       ),
     ]);
+  }
+
+  // perform validation checks on core and bridges
+  async checkDeployment(): Promise<void> {
+    await this.checkCores();
+    await this.checkBridges();
   }
 
   async checkCores(): Promise<void> {
